python: Move the custom_mesh_dot_writer inside m5
This is done so that we can use the new dot_writer
from within stdlib as well
