+++ src/python/SConscript
@@ -432,6 +432,7 @@ PySource('m5.stats', 'm5/stats/__init__.py')
 PySource('m5.util', 'm5/util/__init__.py')
 PySource('m5.util', 'm5/util/attrdict.py')
 PySource('m5.util', 'm5/util/convert.py')
+PySource('m5.util', 'm5/util/custom_mesh_dot_writer.py')
 PySource('m5.util', 'm5/util/dot_writer.py')
 PySource('m5.util', 'm5/util/dot_writer_ruby.py')
 PySource('m5.util', 'm5/util/fdthelper.py')
